Add playGame function
The game needs to be a total of five rounds. Add a function that loops until 5 rounds have been played and console.logs the final score at the end. Call the function via console.log at the end so that users are prompted to play the game upon entering the page. And voila!